test/e2e: Remove incorrect test error
- We were hitting the error:
```
Number of cpus %f specified in annotation default_vcpus is greater than the number of CPUs %d on the system
```
in the tests, due to incorrect behaviour in the remote hypervisor
code in the kata-shim. This should have been fixed in
https://github.com/kata-containers/kata-containers/pull/9657,
so this code path isn't useful anymore and if we hit it, it is a bug, so we want to know.

Signed-off-by: stevenhorsman <steven@uk.ibm.com>