test/e2e: Increase create_container_timeout
We are seeing failures in at least the ibmcloud e2e tests due to the image
pull taking more than 60s even for a fairly simple pod.
This is probably worth looking into, but if we up the create container timeout
to 5mins then it might provide more stability to our tests.

Signed-off-by: stevenhorsman <steven@uk.ibm.com>