test/e2e: Rework PodEventExtractor
I used the `PodEventExtractor` method to compare against an expected
error message, but didn't realise at the time that it just watched for the
first warning message and returned that. We then use this output in
`assessment_runner` to see if our expected error description is there.

Note: As we use it in the `Assess` phase of testing and the pod should
be in the expected state after the `WithSetup`, we should be able to
replace the need for the watch and just get a list of the events

Rename to clarify that only warning (and error) events are considered
and create a `ComparePodEventDescriptions` method in
`assessment_helper` to help simplify the logic in `assessment_runner`
and process all warning events looking for a description match, not just
the first match. If no error was expected then mark the test as failing.

Based on the above logic we can simplify the instance type testing