libvirt: Add the option of EFI firmware support for mkosi
- The packer build doesn't want a custom firmware,
so we can't easily keep the default to be ovmf.
Instead reverse the logic and set it as blank unless provided.
- In order to preserve the existing behaviour set the
kustomization.yaml with the old default and update
the packer e2e testing to comment it out
- Also rename to LIBVIRT_EFI_FIRMWARE to clarify that it's
explicitly used for EFI firmware

- If the EFI firmware path is set, then set domain firmware to efi
and set the device disk to sata & sdb as
mkosi needs sata and packer needs IDE. In future once we
drop packer we can potentially simplify this.

Signed-off-by: stevenhorsman <steven@uk.ibm.com>