caa: Add option to skip Peer pod VM delete
During development, several of us have hit times where being able to
access a failing peer pod VM has been crucial to debugging issues.
The general approach to this is to add a sleep before, or comment out
`DeleteInstance`, but this involved rebuilding the CAA image, so instead
we can provide an option to set `PEERPODS_DEVELOPER_MODE=true`
in peer-pods-cm which will skip this peer pod vm delete.

Signed-off-by: stevenhorsman <steven@uk.ibm.com>